scripts/get_comments/requirements.txt
praw
beautifulsoup4
google-cloud-storage
